# go-echo-backend/services/backend/docs/swagger.yaml
definitions:
  errs.Error:
    properties:
      code:
        type: integer
      header:
        type: integer
      message:
        type: string
    type: object
  location.Coordinate:
    properties:
      address_number:
        type: string
      country_code:
        type: string
      formatted_address:
        type: string
      lat:
        type: number
      level_1:
        type: string
      level_2:
        type: string
      level_3:
        type: string
      level_4:
        type: string
      lng:
        type: number
      postal_code:
        type: string
      street:
        type: string
    type: object
  models.Attachment:
    properties:
      content_type:
        type: string
      file_key:
        type: string
      file_name:
        type: string
      file_url:
        type: string
      metadata:
        additionalProperties: true
        type: object
      thumbnail_url:
        type: string
    required:
    - file_key
    type: object
...
  models.Coordinate:
    properties:
      address_id:
        description: 'Address: just use for query in BE'
        type: string
      address_name:
        type: string
      address_number:
        type: string
      address_type:
        type: string
      country_code:
        type: string
      created_at:
        type: integer
      deleted_at:
        type: integer
      formatted_address:
        type: string
      id:
        type: string
      lat:
        type: number
      level_1:
        type: string
      level_2:
        type: string
      level_3:
        type: string
      level_4:
        type: string
      lng:
        type: number
      postal_code:
        type: string
      street:
        type: string
      updated_at:
        type: integer
    type: object
...
  models.M:
    additionalProperties: true
    type: object
...
  models.User:
    properties:
      account_status:
        type: string
      account_status_changed_at:
        type: integer
      account_verified:
        type: boolean
      account_verified_at:
        type: integer
      annual_revenue:
        type: number
      avatar:
        $ref: '#/definitions/models.Attachment'
        description: Photo
        type: object
      bio:
        type: string
      brand_age:
        type: integer
      brand_name:
        type: string
      coordinate:
        $ref: '#/definitions/models.Coordinate'
        type: object
      coordinate_id:
        type: string
      country_code:
        type: string
      created_at:
        type: integer
      customer_type:
        type: string
      deleted_at:
        type: integer
      description:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_first_login:
        type: boolean
      is_offline:
        type: boolean
      is_test:
        type: boolean
      last_login:
        type: integer
      last_name:
        type: string
      last_online_at:
        type: integer
      monthly_turnover:
        type: number
      name:
        type: string
      phone_number:
        type: string
      register_by_inquiry:
        type: boolean
      representative:
        type: string
      role:
        description: Common
        type: string
      social_provider:
        type: string
      stripe_customer_id:
        type: string
      tax_id:
        type: string
      timezone:
        type: string
      total_pcs_required:
        type: integer
      updated_at:
        type: integer
    type: object
...
  models.UserUpdateForm:
    properties:
      account_status:
        type: string
      account_status_changed_at:
        type: integer
      annual_revenue:
        type: number
      avatar:
        $ref: '#/definitions/models.Attachment'
        type: object
      bio:
        type: string
      brand_age:
        type: integer
      brand_name:
        type: string
      coordinate:
        $ref: '#/definitions/location.Coordinate'
        type: object
      country_code:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      monthly_turnover:
        type: number
      name:
        type: string
      phone_number:
        type: string
      representative:
        type: string
      role:
        description: Common
        type: string
      tax_id:
        type: string
      total_pcs_required:
        type: integer
    type: object
...
  query.Pagination:
    properties:
      current_page:
        type: integer
      has_next:
        type: boolean
      has_prev:
        type: boolean
      metadata:
        type: object
      next_page:
        type: integer
      offset:
        type: integer
      per_page:
        type: integer
      prev_page:
        type: integer
      records:
        type: object
      total_page:
        type: integer
      total_record:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: thaitanloi365@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: inflow api docs
  version: "1.0"
paths:
...
  /api/v1/admin/users:
    get:
      consumes:
      - application/json
      description: Get users
      parameters:
      - default: "1"
        description: Keyword
        in: query
        name: keyword
        type: string
      - default: '"last_login desc"'
        description: Sort
        enum:
        - '"last_login desc"'
        - '"last_login asc"'
        in: query
        items:
          type: string
        name: sorts
        type: array
      - default: 1
        description: Page index
        in: query
        name: page
        type: integer
      - default: 20
        description: Size of page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Bearer:
              description: OK
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/query.Pagination'
            - properties:
                records:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
      security:
      - ApiKeyAuth: []
      summary: Get users
      tags:
      - Admin-User
  /api/v1/admin/users/{user_id}:
    get:
      consumes:
      - application/json
      description: Get user
      parameters:
      - description: ID of user
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Bearer:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
      security:
      - ApiKeyAuth: []
      summary: Get user
      tags:
      - Admin-User
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Form
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UserUpdateForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Bearer:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/models.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Admin-User
...
  /api/v1/admin/users/{user_id}/approve:
    post:
      consumes:
      - application/json
      description: Approve user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
      security:
      - ApiKeyAuth: []
      summary: Approve user
      tags:
      - Admin-User
  /api/v1/admin/users/{user_id}/archive:
    delete:
      consumes:
      - application/json
      description: Unarchive user
      parameters:
      - description: ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Bearer:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/models.M'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
      security:
      - ApiKeyAuth: []
      summary: Unarchive user
      tags:
      - Admin-User
...
  /api/v1/admin/users/{user_id}/reject:
    delete:
      consumes:
      - application/json
      description: Reject user
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
      security:
      - ApiKeyAuth: []
      summary: Reject user
      tags:
      - Admin-User
...
  /api/v1/admin/users/recent:
    get:
      consumes:
      - application/json
      description: Get recent users created
      parameters:
      - default: "1"
        description: Keyword
        in: query
        name: keyword
        type: string
      - default: '"last_login desc"'
        description: Sort
        enum:
        - '"last_login desc"'
        - '"last_login asc"'
        in: query
        items:
          type: string
        name: sorts
        type: array
      - default: 1
        description: Page index
        in: query
        name: page
        type: integer
      - default: 20
        description: Size of page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Bearer:
              description: OK
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/query.Pagination'
            - properties:
                records:
                  items:
                    $ref: '#/definitions/models.User'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
      security:
      - ApiKeyAuth: []
      summary: Get recent users created
      tags:
      - Admin-User
  /api/v1/admin/users/search:
    get:
      consumes:
      - application/json
      description: Search users
      parameters:
      - default: "1"
        description: Keyword
        in: query
        name: keyword
        type: string
      - default: 20
        description: Size of page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Bearer:
              description: OK
              type: string
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.Error'
      security:
      - ApiKeyAuth: []
      summary: Search users
      tags:
      - Admin-User
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"